034-covid-mask-classifier-master/covid-mask-classifier-master/requirements.txt
imutils==0.5.3
joblib==0.16.0
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
matplotlib==3.3.0
onnxruntime==1.4.0
opencv-python==4.2.0.34
Pillow==7.2.0
scikit-learn==0.23.1
skl2onnx==1.7.0
tensorflow-gpu==2.2.0
